Add a new prefix-/postfix-pair of patches to change the algorithm to calculate the military technology level of a merged nation after absorbing another nation.
Instead of weighing the military tech levels of both nations with their number of owned regions, the larger military technology value is used as starting point. If both miltech values are within 0.5 points, the merged nation's miltech value will be increased by up to 0.5 to represent the often different capabilities of military forces with a roughly comparable technological level and the synergy of combining them.
Existing armies and navies might might give a malus to the merged nation's miltech value to represent the need to possibly replace or upgrade equipment and to train personnel.

Yard1 and his mod "Unification Pop Based Miltech Calculation" provided inspiration for this change and I copied the postfix patch from his mod, although I chose a completely different approach for the prefix patch and the actual formula to calculate the merged nation's military tech level. Thank you!