Update AI evaluator patches with recent changes in vanilla
Still not a great situation though.
These patches will require continuous maintenance.
Changing the mod default settings can have unintended consequences on the AI decision making.
Results aren't really identical to vanilla (which appears to have been the original intention).
And it's doubtful if an identical evaluation of a country in vanilla and in this mod would be a great ides, given that IP scaling is vastly different.