More changes to SpoilsEffectsPatch
The magnitude of the public opinion effect of a spoils priority completion now again depends on the global "spoilsPriorityPublicOpinionScaling" variable.
So if the vanilla devs decide to increase or decrease that effect, it will change by the same factor for users of this mod.
This also allows users to test changes themselves again by editing that parameter in TIGlobalConfig.json.

I also increased the public opinion effect by 50% as a quick balance change while there.