Add check for legitimization of hostile claims trigger from vanilla
Start on making the checks for that trigger actually work as intended.

I had originally written a transpiler patch for the vanilla method, but didn't think about the full overwrite patch already in this mod. So this could never work.
This should be  a start to fix this and to later clean up the code added after the release of 0.4.91.