Add overwrite patch for TINationState.MaxAnnualDirectInvestIPs
Change the formula to calculate the maximum annual direct investments in nations.
The number has been increased significantly.
The cap scales linearly with a nation's monthly IP, multiplied with a user-definable factor.
Every nation also gets a flat amount of 200 to allow substantial direct investments in small and poor nations.